requirements.txt
Flask
torch
diffusers
Pillow
transformers
accelerate
scipy
sentencepiece
opencv-python